Refactor to avoid code duplication
