+++ src/tools/generate/generate_sitemap.js
@@ -1,28 +1,17 @@
 const fs = require('fs-extra');
 const ejs = require('ejs');
+const { get_static_lang_year_files } = require('./shared');
+const { get_static_lang_files } = require('./shared');
 
 const min_publish_date = '2019-11-11';
 const sitemap_template = `templates/sitemap.ejs.xml`;
 const sitemap_path = `templates/sitemap.xml`;
-const static_pages_lang_year = [
-  'index.html',
-  'table_of_contents.html',
-  'methodology.html',
-  'contributors.html',
-  'stories/page_content.html',
-  'stories/user_experience.html',
-  'stories/content_publishing.html',
-  'stories/content_distribution.html'
-];
-const static_pages_lang = [
-  'accessibility_statement.html'
-];
 const ebook_path = "static/pdfs/web_almanac_";
 
 const last_updated_json = "config/last_updated.json";
 let file_dates = {};
 
-const generate_sitemap = async (sitemap_chapters,sitemap_languages) => {
+const generate_sitemap = async (sitemap_chapters, sitemap_languages) => {
 
   file_dates = JSON.parse(await fs.readFile(last_updated_json, 'utf8'));
 
@@ -49,24 +38,10 @@ const generate_sitemap = async (sitemap_chapters,sitemap_languages) => {
 
 const get_static_pages = async (sitemap_languages) => {
 
-  var languages_and_years = [];
-  var languages = [];
-
-  for (const year in sitemap_languages) {
-    for (const language in sitemap_languages[year]) {
-      languages_and_years.push(`${sitemap_languages[year][language]}/${year}`);
-      // Get a list of just languages as well
-      const lang_code = `${sitemap_languages[year][language]}`;
-      if (!languages.includes(`${lang_code}`)) languages.push(`${lang_code}`);
-    }
-  }
-
   let urls = [];
 
   // Get all of the static pages for each combination of language and year
-  const files = languages_and_years
-    .map((x) => static_pages_lang_year.map((p) => `${x}/${p}`))
-    .reduce((x, y) => [...x, ...y], []);
+  const files = get_static_lang_year_files(sitemap_languages);
 
   // Get the sitemap entries for those pages
   for (const loc of await files) {
@@ -79,9 +54,7 @@ const get_static_pages = async (sitemap_languages) => {
   }
 
   // Get all of the static pages with no year for each language
-  const files_no_year = languages
-    .map((x) => static_pages_lang.map((p) => `${x}/${p}`))
-    .reduce((x, y) => [...x, ...y], []);
+  const files_no_year = get_static_lang_files(sitemap_languages);
 
   for (const loc of await files_no_year) {
     if (fs.existsSync(`templates/${loc}`)) {
@@ -120,14 +93,7 @@ const convert_file_name = (url) => {
     return url.substr(0, url.length - 10);
   };
   if ( url.endsWith(".html")) {
-    if ( url.endsWith("accessibility_statement.html")) {
-      // Strip year from Accessibility Statement
-      // TODO must fix this properly to avoid clashes
-      // once we know how we'll handle this in future years
-      return url.substr(0, url.length - 5).replace(/_/g,'-').replace(/\/[0-9]{4}/,'');
-    } else {
-      return url.substr(0, url.length - 5).replace(/_/g,'-');
-    }
+    return url.substr(0, url.length - 5).replace(/_/g,'-');
   };
   return url.replace(/_/g,'-');
 };
