+++ src/tools/generate/generate_timestamps.js
@@ -2,22 +2,9 @@ const fs = require('fs-extra');
 const { find_markdown_files } = require('./shared');
 const { find_asset_files } = require('./shared');
 const { get_yearly_configs } = require('./shared');
+const { get_static_lang_year_files } = require('./shared');
+const { get_static_lang_files } = require('./shared');
 const crypto = require('crypto');
-
-const static_pages_lang_year = [
-  'index.html',
-  'table_of_contents.html',
-  'methodology.html',
-  'contributors.html',
-  'ebook.html',
-  'stories/page_content.html',
-  'stories/user_experience.html',
-  'stories/content_publishing.html',
-  'stories/content_distribution.html'
-];
-const static_pages_lang = [
-  'accessibility_statement.html'
-];
 
 const path = "config/last_updated.json";
 
@@ -95,22 +82,8 @@ const get_asset_file_dates = async () => {
 
 const get_template_pages_dates = async (supported_languages) => {
 
-  var languages_and_years = [];
-  var languages = [];
-
-  for (const year in supported_languages) {
-    for (const language in supported_languages[year]) {
-      languages_and_years.push(`${supported_languages[year][language]}/${year}`);
-      // Get a list of just languages as well
-      const lang_code = `${supported_languages[year][language]}`;
-      if (!languages.includes(`${lang_code}`)) languages.push(`${lang_code}`);
-    }
-  }
-
   // Get all of the static pages for each combination
-  const files = languages_and_years
-    .map((x) => static_pages_lang_year.map((p) => `${x}/${p}`))
-    .reduce((x, y) => [...x, ...y], []);
+  const files = get_static_lang_year_files(supported_languages);
 
   // Get the sitemap entries for those pages
   let static_pages_dates = [];
@@ -124,9 +97,7 @@ const get_template_pages_dates = async (supported_languages) => {
   }
 
   // Get all of the static pages with no year for each language
-  const files_no_year = languages
-    .map((x) => static_pages_lang.map((p) => `${x}/${p}`))
-    .reduce((x, y) => [...x, ...y], []);
+  const files_no_year = get_static_lang_files(supported_languages);
 
   for (const file of await files_no_year) {
     if (fs.existsSync(`templates/${file}`)) {
