+++ src/tools/generate/shared.js
@@ -1,5 +1,18 @@
 const fs = require('fs-extra');
 const recursive = require('recursive-readdir');
+const static_pages_lang_year = [
+  'index.html',
+  'table_of_contents.html',
+  'methodology.html',
+  'contributors.html',
+  'stories/page_content.html',
+  'stories/user_experience.html',
+  'stories/content_publishing.html',
+  'stories/content_distribution.html'
+];
+const static_pages_lang = [
+  'accessibility_statement.html'
+];
 
 const find_template_files = async () => {
   const filter = (file, stats) => {
@@ -51,11 +64,11 @@ const get_yearly_configs = async () => {
   let configs = {};
 
   for (const config_file of await find_config_files()) {
-    const re = (process.platform != 'win32') 
-                  ? /config\/([0-9]*).json/ 
+    const re = (process.platform != 'win32')
+                  ? /config\/([0-9]*).json/
                   : /config\\([0-9]*).json/;
     const [path,year] = config_file.match(re);
-    
+
     configs[year] = JSON.parse(await fs.readFile(`config/${year}.json`, 'utf8'));
   }
   return configs;
@@ -82,6 +95,56 @@ const parse_array = (array_as_string) => {
     .map((value) => value.trim()));
 };
 
+const get_languages_and_years_as_array = (language_array) => {
+
+  var languages_and_years = [];
+  var languages = [];
+
+  for (const year in language_array) {
+    for (const language in language_array[year]) {
+      languages_and_years.push(`${language_array[year][language]}/${year}`);
+      // Get a list of just languages as well
+      const lang_code = `${language_array[year][language]}`;
+      if (!languages.includes(`${lang_code}`)) languages.push(`${lang_code}`);
+    }
+  }
+
+  return [languages_and_years,languages];
+
+};
+
+const get_static_lang_year_files = (language_array) => {
+
+  var languages_and_years = [];
+  var languages = [];
+
+  [languages_and_years, languages] = get_languages_and_years_as_array(language_array);
+
+  // Get all of the static pages for each combination of language and year
+  const files = languages_and_years
+    .map((x) => static_pages_lang_year.map((p) => `${x}/${p}`))
+    .reduce((x, y) => [...x, ...y], []);
+
+  return files;
+
+};
+
+const get_static_lang_files = (language_array) => {
+
+  var languages_and_years = [];
+  var languages = [];
+
+  [languages_and_years, languages] = get_languages_and_years_as_array(language_array);
+
+  // Get all of the static pages for each language
+  const files = languages
+    .map((x) => static_pages_lang.map((p) => `${x}/${p}`))
+    .reduce((x, y) => [...x, ...y], []);
+
+  return files;
+
+};
+
 module.exports = {
   find_markdown_files,
   find_template_files,
@@ -89,5 +152,7 @@ module.exports = {
   find_config_files,
   get_yearly_configs,
   size_of,
-  parse_array
+  parse_array,
+  get_static_lang_year_files,
+  get_static_lang_files
 };
